requirements.txt
streamlit==1.38.0
openai==1.51.2
ollama==0.3.3
PyPDF2==3.0.1
pandas==2.2.2
numpy==2.1.1
python-dotenv==1.0.1
httpx==0.27.0
